Initialize tests cleanup
