+++ test/functional/tests/initialize/test_startup_init_config.py
